// src/interfaces/busquedaGeneralPoliza.model.ts
export class BusquedaGeneralPoliza {
    NumeroPoliza: string;
    CUSPP: string;
    TipoDocumento: number;
    NumeroDocumento: string;
    GlsTipoDocumento: string;
    ApellidoPaterno: string;
    ApellidoMaterno: string;
    PrimerNombre: string;
    SegundoNombre: string;
    NumeroEndoso: number;
    GlsParentesco: string;
    Mensaje: string;
    NumOrden: number;
}